スコア送信部分をmainにドッキング
