+++ main.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template
+from flask import Flask, request, render_template
 
 app = Flask(__name__)
 
@@ -6,9 +6,11 @@ app = Flask(__name__)
 def index():
     return render_template('index.html')
 
-@app.route('/game')
-def game():
-    return render_template('game.html')
+@app.route('/receive_data', methods=['POST'])
+def receive_data():
+    data = request.json  # POSTで送信されたJSONデータを受け取る例
+    print("Received data:", data)
+    return "Data received successfully"
 
 if __name__ == '__main__':
     app.run(debug=True, port=5001)  # ポート番号を5001に変更
